# Conflicts solved:
#	source/am_map.cpp
#	source/e_ttypes.cpp
#	source/e_ttypes.h
#	source/p_slopes.cpp
#	source/p_spec.cpp

+++ source/m_buffer.cpp
@@ -384,7 +384,7 @@ size_t InBuffer::read(void *dest, size_t size)
 //
 int InBuffer::skip(size_t skipAmt)
 {
-   return fseek(f, skipAmt, SEEK_CUR);
+   return fseek(f, static_cast<long>(skipAmt), SEEK_CUR);
 }
 
 //
